NISP-2303: Regress personal max & break gapsToFill
The maximum.yearsToWork and maximum.gapsToFill are now broken.
This is because to accoomdate pre16 and post16 gaps to fill, the method of calculating gapsToFill is far more complicated to be implemented within the project's time constraints.
As an interim measure we have taken Personal Max back to just "fill all
the gaps and that is the maximum" without worrying how the user gets
there. The frontend will be updated accordingly.